Update test matrix to include more Ruby versions
Ruby 2.5 was recently released. As Rails 5.1 has released system tests
more people may be going to look towards this gem as a way to proxy
their requests. However, those users will also likely only be using Ruby
2.4+.

In order for this to work in newer Rubies the `watir` needs to be
allowed to run a more current version. But for Ruby 1.9.3 support it
still needs the old version.